Fix naming and convert types.
