+++ core/cli/CMakeLists.txt
@@ -7,7 +7,6 @@ add_library(cli INTERFACE)
 target_link_libraries(cli INTERFACE
     Boost::program_options
     fmt::fmt
-    file
     )
 
 add_subdirectory(node)
