+++ core/cli/node/CMakeLists.txt
@@ -9,6 +9,7 @@ add_executable(fuhon-node-cli
 target_link_libraries(fuhon-node-cli
     cli
     rpc
+    file
     )
 set_target_properties(fuhon-node-cli PROPERTIES
     RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin
