+++ core/cli/node/main.cpp
@@ -20,18 +20,18 @@ namespace fc::cli::_node {
           CMD("peers", Node_net_peers),
       })},
       {GROUP("wallet")({
-          CMD("new", walletNew),
-          CMD("list", walletList),
-          CMD("add-balance", walletAddBalance),
-          CMD("balance", walletBalance),
-          CMD("default", walletDefault),
-          CMD("set-default", walletSetDefault),
-          CMD("import", walletImport),
-          CMD("sign", walletSign),
-          CMD("verify", walletVerify),
-          CMD("delete", walletDelete),
+          CMD("new", Node_wallet_new),
+          CMD("list", Node_wallet_list),
+          CMD("add-balance", Node_wallet_add_balance),
+          CMD("balance", Node_wallet_balance),
+          CMD("default", Node_wallet_default),
+          CMD("set-default", Node_wallet_set_default),
+          CMD("import", Node_wallet_import),
+          CMD("sign", Node_wallet_sign),
+          CMD("verify", Node_wallet_verify),
+          CMD("delete", Node_wallet_delete),
           {GROUP("market")({
-              CMD("add", walletMarketAdd),
+              CMD("add", Node_wallet_market_add),
           })},
       })},
       CMD("version", Node_version),
